# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: Atlas
  version: "1.0.0"
  summary: Auth, geospatial queries, and payments for location-aware apps.
  description: |
    Every `/v1` request carries **two independent credentials**, and
    conflating them is the mistake worth avoiding:

    | Header | Answers | Comes from |
    |---|---|---|
    | `X-Atlas-Key` | which application is calling | `atlas keys create` |
    | `Authorization: Bearer` | which of your users is calling | `POST /v1/auth/login` |

    Neither substitutes for the other, and the project key is required
    even on register and login — creating a user means creating them in a
    project. The project key is a **server-side secret**: anyone holding
    it can act on your whole project, so it must not ship in a browser
    bundle or a mobile app.

    A token is bound to the project that issued it. Presenting one
    alongside a different project's key is a 401.

    **No request body has a `user_id` or `project_id` field.** Both are
    injected by the gateway from the credentials above, so a caller can
    name neither — it can only present credentials that name them.
servers:
  - url: https://api.atlas.dev
    description: Production
  - url: http://localhost:8080
    description: docker-compose
security:
  - projectKey: []
    bearerAuth: []
...
paths:
...
  /v1/payments/wallet:
    get:
      tags: [payments]
      summary: The caller's wallet balance
      responses:
        "200":
          description: |
            Balance in minor units. A user with no wallet reads as zero
            rather than 404 — the wallet is created on first use.
          content:
            application/json:
              schema:
                type: object
                properties:
                  balance_cents: { type: integer, format: int64 }
                  currency: { type: string, example: USD }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "503": { $ref: "#/components/responses/Unavailable" }
  /v1/payments/deposits:
    post:
      tags: [payments]
      summary: Add funds to the caller's wallet
      description: |
        The only way money enters the platform. Settlement refuses to move
        funds a wallet does not hold, so wallets start empty and stay
        honest.

        There is **no real payment processor** behind this yet: the
        provider is a placeholder that always authorises. No money moves.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount_cents]
              properties:
                amount_cents: { type: integer, format: int64, exclusiveMinimum: 0 }
                idempotency_key:
                  type: string
                  description: |
                    Generated for you if omitted. Scoped per project, so
                    two customers may both use `order-1`.
      responses:
        "200":
          description: Deposited.
          content:
            application/json:
              schema:
                type: object
                properties:
                  transaction_id: { type: string, format: uuid }
                  status: { type: string }
                  balance_cents: { type: integer, format: int64 }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/IdempotencyConflict" }
        "503": { $ref: "#/components/responses/Unavailable" }
  /v1/payments/connected-accounts:
    post:
      tags: [payments]
      summary: Start Connect payout onboarding for the caller
      description: |
        Creates (or resumes) the caller's own payout account with the
        project's payment processor and returns the hosted onboarding
        URL. The app must redirect immediately — the URL is single-use
        and short-lived.

        Like deposits, this is safe to expose because the account being
        created is the token subject's: a caller can only onboard
        themselves as a payee.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [return_url]
              properties:
                return_url:
                  type: string
                  description: Absolute URL the processor returns the user to.
      responses:
        "201":
          description: Onboarding started.
          content:
            application/json:
              schema:
                type: object
                properties:
                  onboarding_url: { type: string }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/Unprocessable" }
        "503": { $ref: "#/components/responses/Unavailable" }
    get:
      tags: [payments]
      summary: The caller's payout account status
      description: |
        Whether the caller has connected a payout account and whether
        the processor has finished verifying it. A transaction naming
        this user as payee fails until payouts_enabled is true.
      responses:
        "200":
          description: Status.
          content:
            application/json:
              schema:
                type: object
                properties:
                  exists: { type: boolean }
                  payouts_enabled: { type: boolean }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "503": { $ref: "#/components/responses/Unavailable" }
  /v1/payments/transactions:
    post:
      tags: [payments]
      summary: Move funds from the caller to another user
      description: |
        Creates a **pending** transaction and emits `RIDE_ACCEPTED`. The
        money moves later, when the application publishes `RIDE_COMPLETED`
        or `RIDE_CANCELLED` and fare-consumer settles or refunds.

        There is deliberately no settle or refund endpoint here: the
        gateway cannot verify that the caller owns a transaction.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [to_user_id, amount_cents]
              properties:
                to_user_id: { type: string, format: uuid }
                amount_cents: { type: integer, format: int64, exclusiveMinimum: 0 }
                idempotency_key: { type: string }
                ride_id: { type: string, format: uuid }
      responses:
        "200":
          description: Pending transaction created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  transaction_id: { type: string, format: uuid }
                  status: { type: string, example: pending }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/IdempotencyConflict" }
        "503": { $ref: "#/components/responses/Unavailable" }
components:
  securitySchemes:
    projectKey:
      type: apiKey
      in: header
      name: X-Atlas-Key
      description: |
        Your project key, `atl_live_…` / `atl_test_…` / `atl_dev_…`. A
        server-side secret: anyone holding it can act on your whole
        project.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: An end-user token from `POST /v1/auth/login`.
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable across releases. Branch on this, not on the message.
              enum:
                - invalid_argument
                - unauthenticated
                - permission_denied
                - not_found
                - already_exists
                - failed_precondition
                - rate_limited
                - unavailable
                - internal
            message: { type: string }
...
  responses:
...
    BadRequest:
      description: Malformed or out-of-range input.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: |
        Missing or invalid credentials. Also returned when a token is
        presented with a key from a different project — the message does
        not say which half was wrong, since that would confirm the token
        is otherwise valid.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    IdempotencyConflict:
      description: |
        The key was reused with different arguments. Reusing it with the
        same arguments returns the original result instead.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    Unprocessable:
      description: |
        The request is well-formed but cannot be applied — for example
        onboarding a payee against a deployment whose project has not
        connected a payout processor.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unavailable:
      description: |
        An upstream service could not be reached. The gateway fails
        per-request rather than pulling itself out of service.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }